+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org/)
 since `1.7.0`.
 
 ## [Unreleased]
+
+## [1.16.0-alpha.14] - 2020-10-30
 ### Added
 - Added new algorithm `FOV_SYMMETRIC_SHADOWCAST` based off of the code
   provided [here](https://www.albertford.com/shadowcasting/).